# Demolishable, trivial optimization.