Trait is enabled. Traits seems to be hardly ever have any actions. Can't seem to trigger any action during bot play.

- Avoid creation of enumerator on actions each tick.
- Lookup modifiers once.
- Only remove and clear removeActions if any to remove.

perf.log shows Demolisable taking 10-15+ ms at times. 

```

     4 ms [40425] Trait: Demolishable
   175 ms [41227] Trait: Demolishable
    12 ms [41264] Trait: Demolishable
    17 ms [41349] Trait: Demolishable
    14 ms [42018] Trait: Demolishable
    18 ms [42613] Trait: Demolishable
     9 ms [42959] Trait: Demolishable
     2 ms [43921] Trait: Demolishable
    19 ms [44361] Trait: Demolishable
    15 ms [45922] Trait: Demolishable
    10 ms [46159] Trait: Demolishable
    11 ms [46386] Trait: Demolishable
     2 ms [47359] Trait: Demolishable
    17 ms [47551] Trait: Demolishable
     2 ms [47597] Trait: Demolishable
    14 ms [47674] Trait: Demolishable
    15 ms [48696] Trait: Demolishable
     2 ms [49281] Trait: Demolishable
    15 ms [50097] Trait: Demolishable
     2 ms [50530] Trait: Demolishable
```